src/studentsdb/studentsdb/requirements.txt
Pillow
django-crispy-forms
python-dateutil
